+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changelog
 
 v0.4.0 - TBD
 ---
+ * [#103](https://github.com/basz/SlmLocale/pull/103) UriPathStrategy don't redirect to default locale
 
 v0.3.0
 ---